Update patch set 3
Patch Set 3:

You can try that. I've just been taking performance snapshots while is the very long demo page, typing 100 chars and seeing how long was spent in this function.

TBH after this fix this function ceases to be significant as length=1 is so common, but if your version is as fast then that's fine too.

Patch-set: 3