Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> (New libs should have an entry in AUTHORS.txt.)

This is very-emphatically not a real patch that should ever be merged. It exists entirely as infrastructure for the child commits, which also should never be merged, which are demonstrations of possible approaches to using vue inside VE. As such, I didn't care too much about doing adding-a-library right. 🤩

Patch-set: 1