Improve SourceConverter.getSourceTextFromDataRange performance
The performance of this method is somwething to look at simply
because it loops over every single character in the document.
This adds up quickly when the method is rapidly executed e.g. on
a keyboard event, which is what was proposed in Ic403e0a. While
benchmarking the code in Ic403e0a I found that the bottleneck is
here, accounting for >90% of the runtime of the benchmarked code.

By simply moving the more common check to the top (a typical
document contains way more characters than linebreaks) the
runtime is cut in half.

Bug: T269094
Change-Id: Ie1f3d19ae7c4238e40d8d167f74a155eb6f6de3b