ve.dm.Converter: Fail hard on "about" attributes in weird places
If we find an "about" attribute on a node that is not supposed to be
about-grouped, something is very wrong. Most likely an earlier node
that was supposed to start the about-group has disappeared.

Proceeding in this case will cause content corruption in MW (content
leaking out of template transclusions and extension nodes), so throw
an exception instead.

Bug: T273234
Change-Id: I8ab6bc04077166a7df3ad0aaba16cd7f17fbda4c