Update patch set 2
Patch Set 2:

Setting different background color on code snippets is pretty common though, so doesn't it make sense for VE to account for that, instead of having to do it in every framework that includes it? If nothing overrides the background for <code>, the change is simply a no-op.

Patch-set: 2
CC: Gerrit User 389 <389@e9e9afe9-4712-486d-8885-f54b72dd1951>