Update patch set 2
Patch Set 2:

I did it knowingly as elements with <code> tag already has border(border: 1px solid #eaecf0;) so we must override it otherwise discontinuity in line will not look elegant. Table and figure elements are not written in continuation of text, so it is not necessary for these.

Patch-set: 2