Rebaser: Fix broken text formatting options
Using italic or bold was broken, due to missing dependencies. This
patch tries to fix these to at least add basic.

We tried to pull in as few modules as possible. A lot of dependencies of
the added modules were not added and seems there not even needed.

Since we're mostly on a noob level regarding VE backend internals this is
as far as we got. Might be possible to just have a minimal version of
compareElementsUnannotated() here.

For adding the unicode lib we needed to apply that fix. See the proper
fix upstream  I223f2d276b46a62b639fcb2a27fbc4f118503c3b

Note: Done on a WMDE hack-event, continued at Wikimedia Hackathon 2021

Bug: T283458
Change-Id: Ia132e4d767ec79351177edc5c842220ac3690c6e