Update patch set 2
Patch Set 2:

Yeah, we haven't had the time to solve it properly. The hack where we replace the method from the unicode lib was only necessary because we could not add that lib in a way where the code still works.

I could also imagine a fix where we just c&p the necessary parts from that lib into a custom module.

There are also a few other dependencies not pulled in, but at least in our manual tests we did not run into issues with that.

Patch-set: 2