Update patch set 1
Patch Set 1: Code-Review+2

I suspect that we never wrote any of this with the expectation that someone might unregister some of the foundational context items. 🤔 Making it not-error in that case seems reasonable, and the UX issues it'd cause outside of read-only mode would clearly be the fault of the implementor. 

On which note, @Esanders I wonder if we should also add a check here for it not being read-only mode? (Since we're only showing the alien context to have the delete button, and in read-only that'd inherently not do anything...)

Patch-set: 1
Reviewer: Gerrit User 2915 <2915@e9e9afe9-4712-486d-8885-f54b72dd1951>
Label: Code-Review=+2