Fix incomplete ve.Document.getBranchNodeFromOffset() usage
The method is documented to return null. We need to check for
this, I believe. Otherwise we get a confusing runtime failure.

This patch changes the behavior of ve.ce.Surface.findBlockSlug()
in an edge-case. I believe this is fine. There are 2 private
usages of this method, both in ve.ce.Surface:
* In handleObservedChanges() (line #3481). Here it does not make
  any difference if an error is thrown or null is returned. The
  code does the exact same.
* In onModelSelect() (line #3167). Here the behavior changes
  slightly. The new behavior is to always remove the old
  "focusedBlockSlug", even in case of an error. I think this
  makes sense. This can only happen if selection.getRange() is
  invalid for some reason. If it is, we should not continue to
  act as if it is fine, but reset the (obviously) invalid
  focused state.

Furthermore, this doesn't really change actual behavior. Before
the behavior was unspecified: the code just "crashed".

Bug: T279299
Change-Id: I1cf93a9d5a0f62ae922bed3a8d1854266fa7aa47