Update patch set 2
Patch Set 2:

I mostly agree with the change. But, what if the purpose of these docs was documenting that there is nothing more to be said about these parameters, and that the documentation for them is not simply missing?

Patch-set: 2
Reviewer: Gerrit User 417 <417@e9e9afe9-4712-486d-8885-f54b72dd1951>