Store internal list counter in new document storage and persist
Create a new storage in ve.dm.Document for persistent data.
Use this for ve.dm.InternalList#getNextUniqueNumber. This
means the counter is now tied to the document, instead of
the internal list, which better avoids number clashes when
cloning and merging documents.

Have the auto-save code listen to changes to this object
and store it in session storage.

Bug: T294918
Change-Id: I7ffabd85a230af56134304fdccc6bfd48150d64a