Verify key before removing node from InternalList
This is an upstream commit of a solution to Phabricator feature request
T52568, which aims to allow the modification of reference names (keys).

A limitation of InternalList is that the listKey is not verified whenever
removing a node, leading InternalList.removeNode to remove the firstNode
even if the listKey has been modified and inserted in the place of the
removed (i.e. listKey with former name) listKey. This commit inserts a
check prior to removing the firstNode to see if the name still matches,
and if it does not, performs no operation.

This permits the modification of listKeys and allows for the targeted
feature of renaming reference names in Extension:Cite.

Change-Id: I082ff880d8ea380031402287b1be911c16825e33