getViewportRange: Handle node not being rendered
The CE selection can return null rects if the node is not rendered.
This can happen in block images with certain display modes set to
not render the caption (which still exists in the DM).

In this case go to the edge of the node, and try again, each time
trying the parent.

Bug: T296910
Change-Id: Ie00b30c03bb233b680dbd33af78c0a75e7e2de1a