ve.ui.Surface: Attach mobile context to global overlay
Desktop context only needs to appear on top of the editing surface, so
it's attached to the local overlay. But the mobile context is almost
fullscreen and must appear on top of all other elements, so it should
be attached to the global overlay.

We haven't noticed this issue before, because the VE-MW mobile editor
itself appears on top of all other elements.

Bug: T307849
Change-Id: Ia18f31a299338f94e69f1882e6e477f3a22ae905