Help dialog: Don't register as a documentCommand by default
In full page editors we may want to do this, but a default VE
surface should not only capture events that happen when the target
is focused.

This prevents multiple help dialogs open in multi-target users
of VE such as Flow and DiscussionTools (although DT doesn't do
multiplet targets *yet*).

Bug: T154626
Change-Id: I42e32bd88a2df0f6b34cfc4cce34c522d06d00d4