[WIP] Resolve some out-of-sync issues on Android with Google Keyboard
There are several known issues with Google Keyboard on Android where
certain input sequences result in the CE and DM getting out of sync.
These include:

* Pressing Enter at the end of a branch node, such as a paragraph
* Pressing Backspace with a selection spanning multiple nodes, or
  with a selection that starts at the start of its parent node.

In these cases, the GBoard IME is apparently happy to make DOM changes
in the CE view that are not reflected in the underlying DM, resulting
in a confusing user experience. Since some of the IME events involved
such as deleteContentBackward are uncancellable, attempt to remedy this
by intercepting the relevant events in a beforeinput handler, letting
them proceed, and then fixing up the DM manually after they complete.

TODO:
* Figure out IME tests - I haven't yet managed to generate output
  via eventLogger.html that would reproduce the failure observed on
  the demo page nor verify the fix.
* Finish up Backspace handling.

Bug: T217223
Bug: T202719
Bug: T312558
Change-Id: I9dbcfd00d01aa464101184a2f688f94deb4b73d2