[WIP] Resolve some out-of-sync issues on Android with Google Keyboard
There are several known issues with Google Keyboard on Android where
certain input sequences result in the CE and DM getting out of sync,
such as pressing Enter at the end of a branch node (e.g. a paragraph).
In these cases, the GBoard IME is apparently happy to make DOM changes
in the CE view that are not reflected in the underlying DM, resulting
in a confusing user experience. Since some of the IME events involved
may be uncancellable, attempt to remedy this by intercepting the
relevant events in a beforeinput handler, letting them proceed, and then
fixing up the DM manually after they complete.

TODO:
* Figure out IME tests - I haven't yet managed to generate output
  via eventLogger.html that would reproduce the failure observed on
  the demo page nor verify the fix.

Bug: T312558
Change-Id: I9dbcfd00d01aa464101184a2f688f94deb4b73d2