Fix French translation for message key:"visualeditor-changedesc-unknown"
The original translation was "$1 modifié". The French language uses
gendered nouns, and "modified" also has to be translated according to
the gender of $1 (if $1 is masculine, then it is "$1 modifié" and if $1 is
feminine it is "$1 modifiée"). The sentence was changed from "$1 modifié" to
"Élément $1 modifié" (which means "Element $1 modified") because in this
case, "modified" has to be translated respective to "Élément" (which is
always masculine) and not $1 anymore.

Change-Id: I47aa6f7c958de801e36c8cde5ed69bdc9c39d5b7