Fix incomplete teardown for a ve.init.Target.toolbar reference
I run into this locally when I switched back and forth between the
"view", "edit", and "edit source" tabs. Some code in
ve.init.mw.DesktopArticleTarget.setupToolbarSaveButton that accesses
this.actionsToolbar failed because the object in this.actionsToolbar
was half destroyed. Unfortunately I have no good way to reproduce the
issue. I hope it's obvious that when this.toolbar is destroyed the
stray reference in this.actionsToolbar must be cleared as well.

Change-Id: I3306890f5d6f0fda35d4c43ae7b425f6dd49d8c9