[DNM] Exploring issue with how internalList is merged
Temporarily removing a code path to try to understand why it's
crashing.  There's an assumption that the document fragment will
include the entire internalList of its original parent document, but
this doesn't seem to be true.  We should either make that true maybe
by sharing a ref to the same internalList, or we should fix the code
removed here to work in the new case we've created.

Bug: T336417
Change-Id: If57b89e15b236fe4cfd621240c37f47f8c51e94b