[WIP] Support insertion from unrelated, cloned doc
When the source document has been cloned and therefore has "original"
properties, we assume that the target document is the original and
make some optimizations which only work for that case.

We should only apply this logic if the target document is in fact the
original, otherwise it should be treated as unrelated.

Bug: T336417
Change-Id: If57b89e15b236fe4cfd621240c37f47f8c51e94b