fix(VisualEditor): Adjust VE overlays for dark mode
This patch adjusts JavaScript for Visual Editor overlays for
the Vector skin to improve handling of the dark theme. This
ensures that the overlay does not incorrectly inherit the dark
theme styles until a more permanent solution is implemented.

Changes:
- `ve.ui.Surface.js`: Conditionally added 'notheme' class to
  the global overlay on desktop to prevent the application of
  the dark theme.

This is a temporary workaround and will be removed once the
VisualEditor library is updated to use Codex design variables
and CSS properties.

Bug: T363861
Change-Id: I701cfc932a598d94a3c154be48d707d05a4ab665