Avoid Array.concat
Array.concat creates a new array using extra memory. It can
often be avoided by using:
* Array.push with a spread syntax
* Inline spread syntax when being used with Function.apply

When an unbounded amount of linear data is being pushed,
use ve.batchPush, as browsers have a limit to the number
of arguments that can be passed to a function (usually
around a few hundred thousand).

Change-Id: I9addf43b2857de7bcac6718a54ffe14d2178d528