[WIP] Add '`' as a Sequence to trigger <code> annotations
This is a semi-standard trigger, as used in MarkDown and a few other
contexts, and is extremely unlikely to get in the way of users who
are typing normal content.

WIP – I can't work out how to get the tests to work. In practice it
seems fine, but the test swallows the input?

Change-Id: I124005705961080935097857a52bbf6cf4150f04